[mutable] Restore the layout ID redesign. Fix some related bugs.
- Mesh cloning with flags wasn't replicating the MeshIdPrefix.
- Mesh reformat was sometimes duplicating the system buffers.
- Fix bugs with the wrapping project handling of layout block ids

#jira UE-209986
[REVIEW] [at]pere.rifa
#rnx
#rb pere.rifa

[CL 33543385 by jordi rovira in ue5-main branch]

+++ Engine/Plugins/Experimental/Mutable/Source/MutableTools/Private/MuT/ASTOpImageRasterMesh.h
@@ -24,7 +24,7 @@ namespace mu
 		ASTChild projector;
 
 		/** If layouts are used, this can indicate a single layout block that we want to raster. */
-		int32 BlockId;
+		uint64 BlockId;
 		int8 LayoutIndex;
 
 		/** Size of the image to generate by rasterization of the mesh. */
